+++ finagle-thrift/src/test/scala/com/twitter/finagle/thrift/exp/partitioning/ThriftCustomPartitioningServiceTest.scala
@@ -30,7 +30,7 @@ class ThriftCustomPartitioningServiceTest
         Future.value(idsAndRequests)
     }
 
-    override def responseMergerRegistry: PartitioningStrategy.ResponseMergerRegistry =
+    override val responseMergerRegistry: PartitioningStrategy.ResponseMergerRegistry =
       ResponseMergerRegistry.create.add(AMethod, aResponseMerger)
 
   }